+++ HISTORY.md
@@ -6,6 +6,7 @@
 3. Initialise tkvars in frm_partner
 4. Create history panel with mousewheel support
 5. Implement last used text for each chat mode
+6. Change mode colours key to mode.value
 
 
 ## Version 0.0.8 01 September 2026